requirements.txt
appdirs==1.4.4
arrow==0.16.0
asgiref==3.2.10
attrs==20.2.0
bcrypt==3.2.0
black==20.8b1
blessed==1.17.10
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.3.9
cryptography==3.1.1
dataclasses==0.6
distro==1.5.0
Django==3.1.1
django-common-helpers==0.9.2
django-cron==0.5.1
django-picklefield==3.0.1
djangorestframework==3.11.1
docker==2.7.0
docker-compose==1.14.0rc1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
flake8==3.8.3
idna==2.10
importlib-metadata==2.0.0
jsonschema==2.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
paramiko==2.7.2
pathspec==0.8.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==3.13
regex==2020.7.14
requests==2.11.1
six==1.15.0
sqlparse==0.3.1
texttable==0.8.8
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
websocket-client==0.57.0
zipp==3.2.0
gunicorn==20.0.4